Changed the endpoint url
